move getting playerTypeChoices into the SetupPlayerComponent
