+++ packages/web/pages/profile/setup/playerType.tsx
@@ -2,26 +2,10 @@ import { SetupPlayerType } from 'components/Setup/SetupPlayerType';
 import { SetupProfile } from 'components/Setup/SetupProfile';
 import { SetupContextProvider } from 'contexts/SetupContext';
 import { Player_Type } from 'graphql/autogen/types';
-import { getPlayerTypes } from 'graphql/queries/enums/getPlayerTypes';
 import { useUser } from 'lib/hooks';
-import { InferGetStaticPropsType } from 'next';
 import React, { useState } from 'react';
 
-export const getStaticProps = async () => {
-  const playerTypeChoices = await getPlayerTypes();
-
-  return {
-    props: {
-      playerTypeChoices,
-      hideTopMenu: true,
-    },
-  };
-};
-
-type Props = InferGetStaticPropsType<typeof getStaticProps>;
-
-const PlayerTypeSetup: React.FC<Props> = (props) => {
-  const { playerTypeChoices } = props;
+const PlayerTypeSetup: React.FC = () => {
   const [playerType, setPlayerType] = useState<Player_Type>();
   const { user } = useUser({ redirectTo: '/' });
 
@@ -36,7 +20,6 @@ const PlayerTypeSetup: React.FC<Props> = (props) => {
     <SetupContextProvider>
       <SetupProfile>
         <SetupPlayerType
-          playerTypeChoices={playerTypeChoices}
           playerType={playerType}
           setPlayerType={setPlayerType}
         />
